pyproject.toml
[tool.poetry]
name = "img2train"
version = "0.1.0"
description = ""
authors = ["Shane Blaufuss <79183990+commandline-johnny@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
